+++ crates/tabby-common/src/lib.rs
@@ -6,6 +6,7 @@ pub mod index;
 pub mod languages;
 pub mod path;
 pub mod registry;
+pub mod terminal;
 pub mod usage;
 
 use std::{
@@ -69,30 +70,4 @@ pub struct Package {
 #[derive(Default, Serialize, Deserialize)]
 pub struct DependencyFile {
     pub direct: Vec<Package>,
-}
-
-pub enum HeaderFormat {
-    BoldWhite,
-    BoldBlue,
-}
-
-impl HeaderFormat {
-    fn prefix(&self) -> &str {
-        match self {
-            HeaderFormat::BoldWhite => "\x1b[1m",
-            HeaderFormat::BoldBlue => "\x1b[34;1m",
-        }
-    }
-
-    fn format(&self, header: &str) -> String {
-        format!("{}{header}\x1b[0m", self.prefix())
-    }
-}
-
-pub fn show_info(header: &str, style: HeaderFormat, content: &[&str]) {
-    eprintln!("  {}", style.format(header));
-    for line in content {
-        eprintln!("  {line}");
-    }
-    eprintln!();
 }
