+++ crates/tabby/src/services/model.rs
@@ -3,7 +3,7 @@ use std::{fs, path::PathBuf, sync::Arc};
 use serde::Deserialize;
 use tabby_common::{
     registry::{parse_model_id, ModelRegistry, GGML_MODEL_RELATIVE_PATH},
-    HeaderFormat,
+    terminal::{show_info, HeaderFormat},
 };
 use tabby_download::download_model;
 use tabby_inference::TextGeneration;
@@ -69,12 +69,12 @@ impl PromptInfo {
 
 fn create_ggml_engine(device: &Device, model_path: &str, parallelism: u8) -> impl TextGeneration {
     if !device.ggml_use_gpu() {
-        tabby_common::show_info(
-            "LATENCY",
+        show_info(
+            "CPU Device",
             HeaderFormat::BoldBlue,
             &[
-                "Tabby is running on CPU! Completions will be very slow.",
-                "If you have a supported GPU, please pass --device [cuda/rocm/metal]",
+                "Tabby is currently running on the CPU. Completions may be slow, but it will suffice for testing purposes.",
+                "For better performance, consider deploying Tabby on a GPU device."
             ],
         );
     }
